refactor(shared): add Messenger and MessengerSubscriptionCallback to shared types
